better install scripts

+++ scripts/install_steamcmd.sh
@@ -3,16 +3,13 @@ BASE_DIR="$( cd $SCRIPT_DIR && cd .. && pwd)"
 STEAMCMD_DIR="$BASE_DIR/steamcmd"
 UNAMESTR=`uname`
 
-mkdir -p $STEAMCMD_DIR
-
-if [[ "$UNAMESTR" == 'Linux' ]]; then
+if [ ! -f "$STEAMCMD_DIR/steamcmd.sh" ]; then
   wget "https://steamcdn-a.akamaihd.net/client/installer/steamcmd_linux.tar.gz" -O "$STEAMCMD_DIR/steamcmd.tar.gz"
-elif [[ "$UNAMESTR" == 'Darwin' ]]; then # os x
-  curl -o "$STEAMCMD_DIR/steamcmd.tar.gz" "https://steamcdn-a.akamaihd.net/client/installer/steamcmd_osx.tar.gz"
+  tar -zxvf "$STEAMCMD_DIR/steamcmd.tar.gz" -C $STEAMCMD_DIR
+  rm "$STEAMCMD_DIR/steamcmd.tar.gz"
+  chmod +x "$STEAMCMD_DIR/steamcmd.sh"
+  chmod +x "$STEAMCMD_DIR/steam.sh"
+  echo "successfully installed steamcmd"
+else
+  echo "steamcmd is already installed"
 fi
-
-
-tar -zxvf "$STEAMCMD_DIR/steamcmd.tar.gz" -C $STEAMCMD_DIR
-rm "$STEAMCMD_DIR/steamcmd.tar.gz"
-chmod +x "$STEAMCMD_DIR/steamcmd.sh"
-# chmod +x "$STEAMCMD_DIR/steam.sh"
